+++ app/src/main/java/eu/kanade/tachiyomi/data/download/anime/AnimeDownloader.kt
@@ -5,13 +5,12 @@ import android.content.Context
 import android.content.Intent
 import android.os.Bundle
 import androidx.core.net.toUri
+import com.arthenica.ffmpegkit.FFmpegKit
 import com.arthenica.ffmpegkit.FFmpegKitConfig
-import com.arthenica.ffmpegkit.FFmpegSession
-import com.arthenica.ffmpegkit.FFprobeSession
+import com.arthenica.ffmpegkit.FFprobeKit
 import com.arthenica.ffmpegkit.Level
 import com.arthenica.ffmpegkit.LogCallback
-import com.arthenica.ffmpegkit.ReturnCode
-import com.arthenica.ffmpegkit.SessionState
+import com.arthenica.ffmpegkit.LogRedirectionStrategy
 import com.arthenica.ffmpegkit.StatisticsCallback
 import com.hippo.unifile.UniFile
 import eu.kanade.tachiyomi.animesource.UnmeteredSource
@@ -26,6 +25,7 @@ import eu.kanade.tachiyomi.ui.player.loader.HosterLoader
 import eu.kanade.tachiyomi.util.storage.DiskUtil
 import eu.kanade.tachiyomi.util.storage.toFFmpegString
 import eu.kanade.tachiyomi.util.system.copyToClipboard
+import kotlinx.coroutines.CancellationException
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.Job
@@ -39,15 +39,19 @@ import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.distinctUntilChanged
 import kotlinx.coroutines.flow.filter
 import kotlinx.coroutines.flow.first
+import kotlinx.coroutines.flow.flow
+import kotlinx.coroutines.flow.flowOn
+import kotlinx.coroutines.flow.retryWhen
 import kotlinx.coroutines.flow.transformLatest
 import kotlinx.coroutines.flow.update
-import kotlinx.coroutines.isActive
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.supervisorScope
+import kotlinx.coroutines.suspendCancellableCoroutine
 import logcat.LogPriority
 import tachiyomi.core.common.i18n.stringResource
 import tachiyomi.core.common.storage.extension
 import tachiyomi.core.common.util.lang.launchIO
+import tachiyomi.core.common.util.lang.withIOContext
 import tachiyomi.core.common.util.lang.withUIContext
 import tachiyomi.core.common.util.system.logcat
 import tachiyomi.domain.download.service.DownloadPreferences
@@ -58,8 +62,8 @@ import tachiyomi.i18n.aniyomi.AYMR
 import uy.kohesive.injekt.Injekt
 import uy.kohesive.injekt.api.get
 import uy.kohesive.injekt.injectLazy
-import kotlin.coroutines.cancellation.CancellationException
-import kotlin.coroutines.coroutineContext
+import kotlin.coroutines.resume
+import kotlin.coroutines.resumeWithException
 
 /**
  * This class is the one in charge of downloading episodes.
@@ -112,12 +116,6 @@ class AnimeDownloader(
      */
     val isRunning: Boolean
         get() = downloaderJob?.isActive ?: false
-
-    /**
-     * Whether FFmpeg is running.
-     */
-    @Volatile
-    var isFFmpegRunning: Boolean = false
 
     init {
         scope.launch {
@@ -252,12 +250,10 @@ class AnimeDownloader(
                 removeFromQueue(download)
             }
         } catch (e: Throwable) {
-            if (e is CancellationException) {
-                notifier.onError("Download cancelled")
-            } else {
-                notifier.onError(e.message)
-                logcat(LogPriority.ERROR, e)
-            }
+            if (e is CancellationException) throw e
+            logcat(LogPriority.ERROR, e)
+            notifier.onError(e.message)
+            stop()
         }
     }
 
@@ -265,13 +261,6 @@ class AnimeDownloader(
      * Destroys the downloader subscriptions.
      */
     private fun cancelDownloaderJob() {
-        isFFmpegRunning = false
-        FFmpegKitConfig.getSessions().filter {
-            it.isFFmpeg && (it.state == SessionState.CREATED || it.state == SessionState.RUNNING)
-        }.forEach {
-            it.cancel()
-        }
-
         downloaderJob?.cancel()
         downloaderJob = null
     }
@@ -343,39 +332,64 @@ class AnimeDownloader(
      * @param download the episode to be downloaded.
      */
     private suspend fun downloadEpisode(download: AnimeDownload) {
-        // This try catch manages errors during download
+        val animeDir = provider.getAnimeDir(download.anime.title, download.source)
+
+        val availSpace = DiskUtil.getAvailableStorageSpace(animeDir)
+        if (availSpace != -1L && availSpace < MIN_DISK_SPACE) {
+            download.status = AnimeDownload.State.ERROR
+            notifier.onError(
+                context.stringResource(AYMR.strings.download_insufficient_space),
+                download.episode.name,
+                download.anime.title,
+                download.anime.id,
+            )
+            return
+        }
+
+        val episodeDirname = provider.getEpisodeDirName(download.episode.name, download.episode.scanlator)
+        val tmpDir = animeDir.createDirectory(episodeDirname + TMP_DIR_SUFFIX)!!
+
         try {
-            val animeDir = provider.getAnimeDir(download.anime.title, download.source)
-
-            val availSpace = DiskUtil.getAvailableStorageSpace(animeDir)
-            if (availSpace != -1L && availSpace < MIN_DISK_SPACE) {
-                throw Exception(context.stringResource(AYMR.strings.download_insufficient_space))
-            }
-
-            val episodeDirname = provider.getEpisodeDirName(download.episode.name, download.episode.scanlator)
-            val tmpDir = animeDir.createDirectory(episodeDirname + TMP_DIR_SUFFIX)!!
-
             if (download.video == null) {
                 // Pull video from network and add them to download object
-                try {
-                    val hosters = EpisodeLoader.getHosters(download.episode, download.anime, download.source)
-                    val fetchedVideo = HosterLoader.getBestVideo(download.source, hosters)!!
-
-                    download.video = fetchedVideo
-                } catch (e: Exception) {
-                    logcat(LogPriority.ERROR, e)
+                val hosters = EpisodeLoader.getHosters(download.episode, download.anime, download.source)
+                if (hosters.isEmpty()) {
                     throw Exception(context.stringResource(AYMR.strings.video_list_empty_error))
                 }
+                val bestVideo = HosterLoader.getBestVideo(download.source, hosters)
+                    ?: throw Exception(context.stringResource(AYMR.strings.video_list_empty_error))
+                download.video = bestVideo
             }
 
-            getOrDownloadVideoFile(download, tmpDir)
+            withIOContext {
+                getOrDownloadVideoFile(download, tmpDir)
+            }
 
-            ensureSuccessfulAnimeDownload(download, animeDir, tmpDir, episodeDirname)
-        } catch (e: Exception) {
+            if (!isDownloadSuccessful(download, tmpDir)) {
+                download.status = AnimeDownload.State.ERROR
+                return
+            }
+
+            val filename = DiskUtil.buildValidFilename("${download.anime.title} - ${download.episode.name}")
+            tmpDir.findFile("${filename}_tmp.mkv")?.delete()
+            tmpDir.renameTo(episodeDirname)
+
+            cache.addEpisode(episodeDirname, animeDir, download.anime)
+
+            DiskUtil.createNoMediaFile(tmpDir, context)
+
+            download.status = AnimeDownload.State.DOWNLOADED
+        } catch (error: Throwable) {
+            if (error is CancellationException) throw error
+            // If the video threw, it will resume here
+            logcat(LogPriority.ERROR, error)
             download.status = AnimeDownload.State.ERROR
-            notifier.onError(e.message, download.episode.name, download.anime.title, download.anime.id)
-        } finally {
-            notifier.dismissProgress()
+            notifier.onError(
+                error.message,
+                download.episode.name,
+                download.anime.title,
+                download.anime.id,
+            )
         }
     }
 
@@ -388,7 +402,7 @@ class AnimeDownloader(
     private suspend fun getOrDownloadVideoFile(
         download: AnimeDownload,
         tmpDir: UniFile,
-    ): Video {
+    ) {
         val video = download.video!!
 
         video.status = Video.State.LOAD_VIDEO
@@ -438,16 +452,11 @@ class AnimeDownloader(
             video.status = Video.State.READY
             progressJob?.cancel()
         } catch (e: Exception) {
+            if (e is CancellationException) throw e
             video.status = Video.State.ERROR
             notifier.onError(e.message, download.episode.name, download.anime.title, download.anime.id)
             progressJob?.cancel()
-
-            logcat(LogPriority.ERROR, e)
-
-            throw e
         }
-
-        return video
     }
 
     /**
@@ -462,48 +471,39 @@ class AnimeDownloader(
         tmpDir: UniFile,
         filename: String,
     ): UniFile {
-        var file: UniFile? = null
+        return flow {
+            tmpDir.findFile("$filename.tmp")?.delete()
+            val videoFile = tmpDir.createFile("$filename.tmp")!!
+            try {
+                ffmpegDownload(download, tmpDir, videoFile, filename)
+            } catch (e: Exception) {
+                videoFile.delete()
+                throw e
+            }
 
-        val downloadScope = CoroutineScope(coroutineContext)
-        for (tries in 1..3) {
-            if (downloadScope.isActive) {
-                file = try {
-                    ffmpegDownload(download, tmpDir, filename)
-                } catch (e: Exception) {
-                    notifier.onError(
-                        e.message + ", retrying..",
-                        download.episode.name,
-                        download.anime.title,
-                        download.anime.id,
-                    )
-                    delay(2 * 1000L)
-                    null
+            emit(videoFile)
+        }
+            // Retry 3 times, waiting 2, 4 and 8 seconds between attempts.
+            .retryWhen { _, attempt ->
+                if (attempt < 3) {
+                    delay((2L shl attempt.toInt()) * 1000)
+                    true
+                } else {
+                    false
                 }
             }
-            // If download has been completed successfully we break from retry loop
-            if (file != null) break
-        }
-
-        return if (downloadScope.isActive) {
-            file ?: throw Exception("Downloaded file not found")
-        } else {
-            throw Exception("Download has been stopped")
-        }
+            .flowOn(Dispatchers.IO)
+            .first()
     }
 
     // ffmpeg is always on safe mode
-    private fun ffmpegDownload(
+    private suspend fun ffmpegDownload(
         download: AnimeDownload,
         tmpDir: UniFile,
+        videoFile: UniFile,
         filename: String,
-    ): UniFile {
+    ) {
         val video = download.video!!
-
-        isFFmpegRunning = true
-
-        // always delete tmp file
-        tmpDir.findFile("$filename.tmp")?.delete()
-        val videoFile = tmpDir.createFile("$filename.tmp")!!
 
         val ffmpegFilename = { videoFile.uri.toFFmpegString(context) }
 
@@ -512,6 +512,7 @@ class AnimeDownloader(
             "${it.first}: ${it.second}\r\n"
         }
 
+        FFmpegKitConfig.setLogRedirectionStrategy(LogRedirectionStrategy.ALWAYS_PRINT_LOGS)
         val ffmpegOptions = getFFmpegOptions(video, headerOptions, ffmpegFilename())
         val ffprobeCommand = { file: String, ffprobeHeaders: String? ->
             FFmpegKitConfig.parseArguments(
@@ -538,27 +539,27 @@ class AnimeDownloader(
             }
         }
 
-        val session = FFmpegSession.create(ffmpegOptions, {}, logCallback, statCallback)
-        val inputDuration = getDuration(ffprobeCommand(video.videoUrl, headerOptions)) ?: 0F
+        duration = getDuration(ffprobeCommand(video.videoUrl, headerOptions))?.toLong() ?: 0L
 
-        duration = inputDuration.toLong()
-
-        if (!isFFmpegRunning) {
-            throw Exception("ffmpeg was cancelled")
-        }
-        FFmpegKitConfig.ffmpegExecute(session)
-
-        return if (ReturnCode.isSuccess(session.returnCode)) {
-            val file = tmpDir.findFile("$filename.tmp")?.apply {
-                renameTo("$filename.mkv")
+        suspendCancellableCoroutine { continuation ->
+            val session = FFmpegKit.executeWithArgumentsAsync(
+                ffmpegOptions,
+                {
+                    if (it.returnCode.isValueSuccess) {
+                        tmpDir.findFile("$filename.tmp")?.apply {
+                            renameTo("$filename.mkv")
+                        }
+                        continuation.resume(it)
+                    } else {
+                        continuation.resumeWithException(Exception("Error in ffmpeg!"))
+                    }
+                },
+                logCallback,
+                statCallback,
+            )
+            continuation.invokeOnCancellation {
+                session.cancel()
             }
-
-            file ?: throw Exception("Downloaded file not found")
-        } else {
-            session.failStackTrace?.let { trace ->
-                logcat(LogPriority.ERROR) { trace }
-            }
-            throw Exception("Error in ffmpeg!")
         }
     }
 
@@ -618,10 +619,17 @@ class AnimeDownloader(
         return FFmpegKitConfig.parseArguments(command)
     }
 
-    private fun getDuration(ffprobeCommand: Array<String>): Float? {
-        val session = FFprobeSession.create(ffprobeCommand)
-        FFmpegKitConfig.ffprobeExecute(session)
-        return session.allLogsAsString.trim().toFloatOrNull()
+    private suspend fun getDuration(ffprobeCommand: Array<String>): Float? {
+        return suspendCancellableCoroutine { continuation ->
+            val session = FFprobeKit.executeWithArgumentsAsync(ffprobeCommand) {
+                if (it.returnCode.isValueSuccess) {
+                    continuation.resume(it)
+                } else {
+                    continuation.resumeWithException(Exception(it.output))
+                }
+            }
+            continuation.invokeOnCancellation { session.cancel() }
+        }.output.toFloatOrNull()
     }
 
     /**
@@ -726,6 +734,20 @@ class AnimeDownloader(
             tmpDir.findFile("${filename}_tmp.mkv")?.delete()
             throw e
         }
+    }
+
+    /**
+     * Checks if the download was successful.
+     *
+     * @param download the download to check.
+     * @param tmpDir the directory where the download is currently stored.
+     */
+    private fun isDownloadSuccessful(
+        download: AnimeDownload,
+        tmpDir: UniFile,
+    ): Boolean {
+        val downloadedVideo = tmpDir.listFiles().orEmpty().filterNot { it.extension == ".tmp" }
+        return downloadedVideo.size == 1
     }
 
     /**